refactor: multiple Settings updates
