+++ src/components/settings/console/ConsoleSettings.vue
@@ -25,9 +25,9 @@
         </app-btn>
       </app-setting>
 
-      <v-divider />
-
       <template v-for="filter in filters">
+        <v-divider :key="`divider-${filter.name}`" />
+
         <app-setting
           :key="`filter-${filter.name}`"
           :r-cols="3"
@@ -60,8 +60,6 @@
             </v-icon>
           </app-btn>
         </app-setting>
-
-        <v-divider :key="`divider-${filter.name}`" />
       </template>
 
       <console-filter-dialog
